fix: recent envelopes for Accounts in v1, links for Accounts in v2
This refactors the Account code by assembling the API response in
the controller instead of the models.

This enabled fixing the links for Accounts in API v2. Additionally,
the implementation for recent Envelopes already used in API v2 is
now also used in API v1.

This resolves a bug in API v1, where the recent Envelopes did factor in
all transactions for the account, not only recent transactions.
