+++ pkg/controllers/account_v2.go
@@ -1,14 +1,80 @@
 package controllers
 
 import (
+	"fmt"
 	"net/http"
 
+	"github.com/envelope-zero/backend/v3/internal/types"
+	"github.com/envelope-zero/backend/v3/pkg/database"
 	"github.com/envelope-zero/backend/v3/pkg/httperrors"
 	"github.com/envelope-zero/backend/v3/pkg/httputil"
 	"github.com/envelope-zero/backend/v3/pkg/models"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
+	"github.com/shopspring/decimal"
 )
+
+// AccountV2 is the API v2 representation of an Account in EZ.
+type AccountV2 struct {
+	models.Account
+	Balance           decimal.Decimal `json:"balance" example:"2735.17"`           // Balance of the account, including all transactions referencing it
+	ReconciledBalance decimal.Decimal `json:"reconciledBalance" example:"2539.57"` // Balance of the account, including all reconciled transactions referencing it
+	RecentEnvelopes   []*uuid.UUID    `json:"recentEnvelopes"`                     // Envelopes recently used with this account
+
+	Links struct {
+		Self         string `json:"self" example:"https://example.com/api/v2/accounts/af892e10-7e0a-4fb8-b1bc-4b6d88401ed2"`                     // The account itself
+		Transactions string `json:"transactions" example:"https://example.com/api/v2/transactions?account=af892e10-7e0a-4fb8-b1bc-4b6d88401ed2"` // Transactions referencing the account
+	} `json:"links"`
+}
+
+// links generates the HATEOAS links for the Account.
+func (a *AccountV2) links(c *gin.Context) {
+	url := c.GetString(string(database.ContextURL))
+	a.Links.Self = fmt.Sprintf("%s/v2/accounts/%s", url, a.ID)
+	a.Links.Transactions = fmt.Sprintf("%s/v2/transactions?account=%s", url, a.ID)
+}
+
+func (co Controller) getAccountV2(c *gin.Context, id uuid.UUID) (AccountV2, bool) {
+	accountModel, ok := getResourceByIDAndHandleErrors[models.Account](c, co, id)
+
+	account := AccountV2{
+		Account: accountModel,
+	}
+
+	if !ok {
+		return AccountV2{}, false
+	}
+
+	// Recent Envelopes
+	ids, err := accountModel.RecentEnvelopes(co.DB)
+	if err != nil {
+		httperrors.Handler(c, err)
+		return AccountV2{}, false
+	}
+
+	account.RecentEnvelopes = ids
+
+	// Balance
+	balance, _, err := accountModel.GetBalanceMonth(co.DB, types.Month{})
+	if err != nil {
+		httperrors.Handler(c, err)
+		return AccountV2{}, false
+	}
+	account.Balance = balance
+
+	// Reconciled Balance
+	reconciledBalance, err := accountModel.SumReconciled(co.DB)
+	if err != nil {
+		httperrors.Handler(c, err)
+		return AccountV2{}, false
+	}
+	account.ReconciledBalance = reconciledBalance
+
+	// Links
+	account.links(c)
+
+	return account, true
+}
 
 // RegisterAccountRoutes registers the routes for accounts with
 // the RouterGroup that is passed.
@@ -67,13 +133,13 @@ func (co Controller) GetAccountsV2(c *gin.Context) {
 		return
 	}
 
-	query := co.DB.Where(&models.AccountV2{
+	query := co.DB.Where(&models.Account{
 		AccountCreate: create,
 	}, queryFields...)
 
 	query = stringFilters(co.DB, query, setFields, filter.Name, filter.Note, filter.Search)
 
-	var accounts []models.AccountV2
+	var accounts []models.Account
 	if !queryWithRetry(c, query.Find(&accounts)) {
 		return
 	}
@@ -81,35 +147,12 @@ func (co Controller) GetAccountsV2(c *gin.Context) {
 	// When there are no resources, we want an empty list, not null
 	// Therefore, we use make to create a slice with zero elements
 	// which will be marshalled to an empty JSON array
-	accountObjects := make([]models.AccountV2, 0)
+	accountObjects := make([]AccountV2, 0)
 
 	for _, account := range accounts {
-		o, _ := co.getAccountObjectV2(c, account.ID)
+		o, _ := co.getAccountV2(c, account.ID)
 		accountObjects = append(accountObjects, o)
 	}
 
 	c.JSON(http.StatusOK, accountObjects)
-}
-
-// getAccountObjectV2 returns the account object with all calculations done.
-func (co Controller) getAccountObjectV2(c *gin.Context, id uuid.UUID) (models.AccountV2, bool) {
-	account, ok := getResourceByIDAndHandleErrors[models.AccountV2](c, co, id)
-
-	if !ok {
-		return models.AccountV2{}, false
-	}
-
-	err := account.SetRecentEnvelopes(co.DB)
-	if err != nil {
-		httperrors.Handler(c, err)
-		return models.AccountV2{}, false
-	}
-
-	err = account.WithCalculations(co.DB)
-	if err != nil {
-		httperrors.Handler(c, err)
-		return models.AccountV2{}, false
-	}
-
-	return account, true
 }
