+++ pkg/models/account_test.go
@@ -11,27 +11,11 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func (suite *TestSuiteStandard) createTestAccountV2(c models.AccountCreate) models.AccountV2 {
-	if c.Name == "" {
-		c.Name = uuid.New().String()
-	}
-
-	account := models.AccountV2{
-		AccountCreate: c,
-	}
-	err := suite.db.Save(&account).Error
-	if err != nil {
-		suite.Assert().FailNow("Account could not be saved", "Error: %s, Account: %#v", err, account)
-	}
-
-	return account
-}
-
-func (suite *TestSuiteStandard) TestAccountV2Calculations() {
+func (suite *TestSuiteStandard) TestAccountCalculations() {
 	budget := suite.createTestBudget(models.BudgetCreate{})
 	initialBalanceDate := time.Now()
 
-	account := suite.createTestAccountV2(models.AccountCreate{
+	account := suite.createTestAccount(models.AccountCreate{
 		Name:               "TestAccountCalculations",
 		BudgetID:           budget.ID,
 		OnBudget:           true,
@@ -40,7 +24,7 @@ func (suite *TestSuiteStandard) TestAccountV2Calculations() {
 		InitialBalanceDate: &initialBalanceDate,
 	})
 
-	externalAccount := suite.createTestAccountV2(models.AccountCreate{
+	externalAccount := suite.createTestAccount(models.AccountCreate{
 		BudgetID: budget.ID,
 		External: true,
 	})
@@ -80,14 +64,17 @@ func (suite *TestSuiteStandard) TestAccountV2Calculations() {
 		Note:                 "Future Income Transaction",
 	})
 
-	err := account.WithCalculations(suite.db)
+	balance, _, err := account.GetBalanceMonth(suite.db, types.Month{})
+	assert.Nil(suite.T(), err)
+
+	reconciled, err := account.SumReconciled(suite.db)
 	assert.Nil(suite.T(), err)
 
 	expected := incomingTransaction.Amount.Sub(outgoingTransaction.Amount).Add(account.InitialBalance).Add(decimal.NewFromFloat(100)) // Add 100 for futureIncomeTransaction
-	assert.True(suite.T(), account.Balance.Equal(expected), "Balance for account is not correct. Should be: %v but is %v", expected, account.Balance)
+	assert.True(suite.T(), balance.Equal(expected), "Balance for account is not correct. Should be: %v but is %v", expected, balance)
 
 	expected = incomingTransaction.Amount.Add(account.InitialBalance)
-	assert.True(suite.T(), account.ReconciledBalance.Equal(expected), "Reconciled balance for account is not correct. Should be: %v but is %v", expected, account.ReconciledBalance)
+	assert.True(suite.T(), reconciled.Equal(expected), "Reconciled balance for account is not correct. Should be: %v but is %v", expected, reconciled)
 
 	balanceNow, availableNow, err := account.GetBalanceMonth(suite.db, types.MonthOf(time.Now()))
 	assert.Nil(suite.T(), err)
@@ -103,24 +90,27 @@ func (suite *TestSuiteStandard) TestAccountV2Calculations() {
 		suite.Assert().Fail("Resource could not be deleted", err)
 	}
 
-	err = account.WithCalculations(suite.db)
+	balance, _, err = account.GetBalanceMonth(suite.db, types.Month{})
+	assert.Nil(suite.T(), err)
+
+	reconciled, err = account.SumReconciled(suite.db)
 	assert.Nil(suite.T(), err)
 
 	expected = outgoingTransaction.Amount.Neg().Add(account.InitialBalance).Add(decimal.NewFromFloat(100)) // Add 100 for futureIncomeTransaction
-	assert.True(suite.T(), account.Balance.Equal(expected), "Balance for account is not correct. Should be: %v but is %v", expected, account.Balance)
+	assert.True(suite.T(), balance.Equal(expected), "Balance for account is not correct. Should be: %v but is %v", expected, balance)
 
 	expected = decimal.NewFromFloat(0).Add(account.InitialBalance)
-	assert.True(suite.T(), account.ReconciledBalance.Equal(expected), "Reconciled balance for account is not correct. Should be: %v but is %v", expected, account.ReconciledBalance)
+	assert.True(suite.T(), reconciled.Equal(expected), "Reconciled balance for account is not correct. Should be: %v but is %v", expected, reconciled)
 }
 
-func (suite *TestSuiteStandard) TestAccountV2Transactions() {
+func (suite *TestSuiteStandard) TestAccountTransactions() {
 	account := models.Account{}
 
 	transactions := account.Transactions(suite.db)
 	assert.Len(suite.T(), transactions, 0)
 }
 
-func (suite *TestSuiteStandard) TestAccountV2OnBudget() {
+func (suite *TestSuiteStandard) TestAccountOnBudget() {
 	account := models.Account{
 		AccountCreate: models.AccountCreate{
 			OnBudget: true,
@@ -150,10 +140,145 @@ func (suite *TestSuiteStandard) TestAccountV2OnBudget() {
 	assert.True(suite.T(), account.OnBudget, "OnBudget is set to false even though the account is internal")
 }
 
-func (suite *TestSuiteStandard) TestAccountV2RecentEnvelopes() {
+func (suite *TestSuiteStandard) TestAccountGetBalanceMonthDBFail() {
+	account := models.Account{}
+
+	suite.CloseDB()
+
+	_, _, err := account.GetBalanceMonth(suite.db, types.NewMonth(2017, 7))
+	suite.Assert().NotNil(err)
+	suite.Assert().Equal("sql: database is closed", err.Error())
+}
+
+// TestAccountDuplicateNames ensures that two accounts cannot have the same name.
+func (suite *TestSuiteStandard) TestAccountDuplicateNames() {
 	budget := suite.createTestBudget(models.BudgetCreate{})
 
-	account := suite.createTestAccountV2(models.AccountCreate{
+	_ = suite.createTestAccount(models.AccountCreate{
+		BudgetID: budget.ID,
+		Name:     "TestAccountDuplicateNames",
+	})
+
+	externalAccount := models.Account{
+		AccountCreate: models.AccountCreate{
+			BudgetID: budget.ID,
+			Name:     "TestAccountDuplicateNames",
+			External: true,
+		},
+	}
+	err := suite.db.Save(&externalAccount).Error
+	if err == nil {
+		suite.Assert().Fail("Account with the same name than another account could be saved. This must not be possible", err)
+		return
+	}
+
+	suite.Assert().Contains(err.Error(), "UNIQUE constraint failed: accounts.name, accounts.budget_id", "Error message for account creation fail does not match expected message")
+}
+
+func (suite *TestSuiteStandard) TestAccountOnBudgetToOnBudgetTransactionsNoEnvelopes() {
+	budget := suite.createTestBudget(models.BudgetCreate{
+		Name: "TestAccountOnBudgetToOnBudgetTransactionsNoEnvelopes",
+	})
+
+	account := suite.createTestAccount(models.AccountCreate{
+		BudgetID: budget.ID,
+		OnBudget: true,
+		External: false,
+		Name:     "TestAccountOnBudgetToOnBudgetTransactionsNoEnvelopes",
+	})
+
+	transferTargetAccount := suite.createTestAccount(models.AccountCreate{
+		BudgetID: budget.ID,
+		OnBudget: false,
+		External: false,
+		Name:     "TestAccountOnBudgetToOnBudgetTransactionsNoEnvelopes:Target",
+	})
+
+	category := suite.createTestCategory(models.CategoryCreate{
+		BudgetID: budget.ID,
+	})
+
+	envelope := suite.createTestEnvelope(models.EnvelopeCreate{
+		CategoryID: category.ID,
+	})
+
+	t := suite.createTestTransaction(models.TransactionCreate{
+		Amount:               decimal.NewFromFloat(17.23),
+		BudgetID:             budget.ID,
+		SourceAccountID:      account.ID,
+		DestinationAccountID: transferTargetAccount.ID,
+		EnvelopeID:           &envelope.ID,
+	})
+
+	// Try saving the account, which must fail
+	data := models.Account{AccountCreate: models.AccountCreate{OnBudget: true}}
+	err := suite.db.Model(&transferTargetAccount).Select("OnBudget").Updates(data).Error
+
+	if !assert.NotNil(suite.T(), err, "Target account could be updated to be on budget while having transactions with envelopes being set") {
+		assert.FailNow(suite.T(), "Exiting because assertion was not met")
+	}
+	assert.Contains(suite.T(), err.Error(), "the account cannot be set to on budget because the following transactions have an envelope set: ")
+
+	// Update the envelope for the transaction
+	t.EnvelopeID = nil
+	err = suite.db.Model(&t).Updates(&t).Error
+	assert.Nil(suite.T(), err, "Transaction could not be updated")
+
+	// Save again
+	err = suite.db.Model(&transferTargetAccount).Updates(&transferTargetAccount).Error
+	assert.Nil(suite.T(), err, "Target account could not be updated despite transaction having its envelope removed")
+}
+
+func (suite *TestSuiteStandard) TestAccountOffBudgetToOnBudgetTransactionsNoEnvelopes() {
+	budget := suite.createTestBudget(models.BudgetCreate{
+		Name: "TestAccountOffBudgetToOnBudgetTransactionsNoEnvelopes",
+	})
+
+	account := suite.createTestAccount(models.AccountCreate{
+		BudgetID: budget.ID,
+		OnBudget: false,
+		External: false,
+		Name:     "TestAccountOffBudgetToOnBudgetTransactionsNoEnvelopes",
+	})
+
+	transferTargetAccount := suite.createTestAccount(models.AccountCreate{
+		BudgetID: budget.ID,
+		OnBudget: false,
+		External: false,
+		Name:     "TestAccountOffBudgetToOnBudgetTransactionsNoEnvelopes:Target",
+	})
+
+	category := suite.createTestCategory(models.CategoryCreate{
+		BudgetID: budget.ID,
+	})
+
+	envelope := suite.createTestEnvelope(models.EnvelopeCreate{
+		CategoryID: category.ID,
+	})
+
+	_ = suite.createTestTransaction(models.TransactionCreate{
+		Amount:               decimal.NewFromFloat(17.23),
+		BudgetID:             budget.ID,
+		SourceAccountID:      account.ID,
+		DestinationAccountID: transferTargetAccount.ID,
+		EnvelopeID:           &envelope.ID,
+	})
+
+	// Try saving the account, which must work
+	data := models.Account{AccountCreate: models.AccountCreate{OnBudget: true}}
+	err := suite.db.Model(&transferTargetAccount).Select("OnBudget").Updates(data).Error
+
+	assert.Nil(suite.T(), err, "Target account could not be updated to be on budget, but it does not have transactions with envelopes being set")
+}
+
+func (suite *TestSuiteStandard) TestAccountSelf() {
+	assert.Equal(suite.T(), "Account", models.Account{}.Self())
+}
+
+func (suite *TestSuiteStandard) TestAccountRecentEnvelopes() {
+	budget := suite.createTestBudget(models.BudgetCreate{})
+
+	account := suite.createTestAccount(models.AccountCreate{
 		BudgetID:       budget.ID,
 		Name:           "Internal Account",
 		OnBudget:       true,
@@ -161,7 +286,7 @@ func (suite *TestSuiteStandard) TestAccountV2RecentEnvelopes() {
 		InitialBalance: decimal.NewFromFloat(170),
 	})
 
-	externalAccount := suite.createTestAccountV2(models.AccountCreate{
+	externalAccount := suite.createTestAccount(models.AccountCreate{
 		BudgetID: budget.ID,
 		Name:     "External Account",
 		External: true,
@@ -213,157 +338,22 @@ func (suite *TestSuiteStandard) TestAccountV2RecentEnvelopes() {
 		Amount:               decimal.NewFromFloat(1337.42),
 	})
 
-	err := account.SetRecentEnvelopes(suite.db)
+	ids, err := account.RecentEnvelopes(suite.db)
 	if err != nil {
 		suite.Assert().FailNow("Could not compute recent envelopes", err)
 	}
 
-	if !suite.Assert().Len(account.RecentEnvelopes, 4, "The number of envelopes in recentEnvelopes is not correct, expected 4, got %d", len(account.RecentEnvelopes)) {
+	if !suite.Assert().Len(ids, 4, "The number of envelopes in recentEnvelopes is not correct, expected 4, got %d", len(ids)) {
 		suite.FailNow("Incorrect envelope number")
 	}
 
 	// The last envelope needs to be the first in the sort since it
 	// has been the most common one
-	suite.Assert().Equal(envelopeIDs[2], account.RecentEnvelopes[0])
+	suite.Assert().Equal(envelopeIDs[2], ids[0])
 
 	// Order for envelopes with the same frequency is undefined
 
 	// Income is the last one since it only appears once
 	var nilUUIDPointer *uuid.UUID
-	suite.Assert().Equal(nilUUIDPointer, account.RecentEnvelopes[3])
-}
-
-func (suite *TestSuiteStandard) TestAccountV2GetBalanceMonthDBFail() {
-	account := models.Account{}
-
-	suite.CloseDB()
-
-	_, _, err := account.GetBalanceMonth(suite.db, types.NewMonth(2017, 7))
-	suite.Assert().NotNil(err)
-	suite.Assert().Equal("sql: database is closed", err.Error())
-}
-
-// TestAccountDuplicateNames ensures that two accounts cannot have the same name.
-func (suite *TestSuiteStandard) TestAccountV2DuplicateNames() {
-	budget := suite.createTestBudget(models.BudgetCreate{})
-
-	_ = suite.createTestAccountV2(models.AccountCreate{
-		BudgetID: budget.ID,
-		Name:     "TestAccountDuplicateNames",
-	})
-
-	externalAccount := models.Account{
-		AccountCreate: models.AccountCreate{
-			BudgetID: budget.ID,
-			Name:     "TestAccountDuplicateNames",
-			External: true,
-		},
-	}
-	err := suite.db.Save(&externalAccount).Error
-	if err == nil {
-		suite.Assert().Fail("Account with the same name than another account could be saved. This must not be possible", err)
-		return
-	}
-
-	suite.Assert().Contains(err.Error(), "UNIQUE constraint failed: accounts.name, accounts.budget_id", "Error message for account creation fail does not match expected message")
-}
-
-func (suite *TestSuiteStandard) TestAccountV2OnBudgetToOnBudgetTransactionsNoEnvelopes() {
-	budget := suite.createTestBudget(models.BudgetCreate{
-		Name: "TestAccountOnBudgetToOnBudgetTransactionsNoEnvelopes",
-	})
-
-	account := suite.createTestAccountV2(models.AccountCreate{
-		BudgetID: budget.ID,
-		OnBudget: true,
-		External: false,
-		Name:     "TestAccountOnBudgetToOnBudgetTransactionsNoEnvelopes",
-	})
-
-	transferTargetAccount := suite.createTestAccountV2(models.AccountCreate{
-		BudgetID: budget.ID,
-		OnBudget: false,
-		External: false,
-		Name:     "TestAccountOnBudgetToOnBudgetTransactionsNoEnvelopes:Target",
-	})
-
-	category := suite.createTestCategory(models.CategoryCreate{
-		BudgetID: budget.ID,
-	})
-
-	envelope := suite.createTestEnvelope(models.EnvelopeCreate{
-		CategoryID: category.ID,
-	})
-
-	t := suite.createTestTransaction(models.TransactionCreate{
-		Amount:               decimal.NewFromFloat(17.23),
-		BudgetID:             budget.ID,
-		SourceAccountID:      account.ID,
-		DestinationAccountID: transferTargetAccount.ID,
-		EnvelopeID:           &envelope.ID,
-	})
-
-	// Try saving the account, which must fail
-	data := models.Account{AccountCreate: models.AccountCreate{OnBudget: true}}
-	err := suite.db.Model(&transferTargetAccount).Select("OnBudget").Updates(data).Error
-
-	if !assert.NotNil(suite.T(), err, "Target account could be updated to be on budget while having transactions with envelopes being set") {
-		assert.FailNow(suite.T(), "Exiting because assertion was not met")
-	}
-	assert.Contains(suite.T(), err.Error(), "the account cannot be set to on budget because the following transactions have an envelope set: ")
-
-	// Update the envelope for the transaction
-	t.EnvelopeID = nil
-	err = suite.db.Model(&t).Updates(&t).Error
-	assert.Nil(suite.T(), err, "Transaction could not be updated")
-
-	// Save again
-	err = suite.db.Model(&transferTargetAccount).Updates(&transferTargetAccount).Error
-	assert.Nil(suite.T(), err, "Target account could not be updated despite transaction having its envelope removed")
-}
-
-func (suite *TestSuiteStandard) TestAccountV2OffBudgetToOnBudgetTransactionsNoEnvelopes() {
-	budget := suite.createTestBudget(models.BudgetCreate{
-		Name: "TestAccountOffBudgetToOnBudgetTransactionsNoEnvelopes",
-	})
-
-	account := suite.createTestAccountV2(models.AccountCreate{
-		BudgetID: budget.ID,
-		OnBudget: false,
-		External: false,
-		Name:     "TestAccountOffBudgetToOnBudgetTransactionsNoEnvelopes",
-	})
-
-	transferTargetAccount := suite.createTestAccountV2(models.AccountCreate{
-		BudgetID: budget.ID,
-		OnBudget: false,
-		External: false,
-		Name:     "TestAccountOffBudgetToOnBudgetTransactionsNoEnvelopes:Target",
-	})
-
-	category := suite.createTestCategory(models.CategoryCreate{
-		BudgetID: budget.ID,
-	})
-
-	envelope := suite.createTestEnvelope(models.EnvelopeCreate{
-		CategoryID: category.ID,
-	})
-
-	_ = suite.createTestTransaction(models.TransactionCreate{
-		Amount:               decimal.NewFromFloat(17.23),
-		BudgetID:             budget.ID,
-		SourceAccountID:      account.ID,
-		DestinationAccountID: transferTargetAccount.ID,
-		EnvelopeID:           &envelope.ID,
-	})
-
-	// Try saving the account, which must work
-	data := models.Account{AccountCreate: models.AccountCreate{OnBudget: true}}
-	err := suite.db.Model(&transferTargetAccount).Select("OnBudget").Updates(data).Error
-
-	assert.Nil(suite.T(), err, "Target account could not be updated to be on budget, but it does not have transactions with envelopes being set")
-}
-
-func (suite *TestSuiteStandard) TestAccountV2Self() {
-	assert.Equal(suite.T(), "Account", models.Account{}.Self())
+	suite.Assert().Equal(nilUUIDPointer, ids[3])
 }
